update: img rotation will now pause and resume according to music is playing, move step function from MusicPlayer.tsx to visualizer.tsx
