+++ src/layouts/MusicPlayer.tsx
@@ -18,8 +18,6 @@ const MusicPlayer = () => {
   const [audioUrl, setAudioUrl] = useState("");
   const audioRef = useRef<HTMLAudioElement | null>(null);
   const musicData = playingQueue?.songs || "";
-  const rafRef = useRef<number>();
-  let lastTime: number;
 
   const handlePlayEnd = () => {
     if (!playingQueue) return;
@@ -101,13 +99,11 @@ const MusicPlayer = () => {
         audioContext.resume();
       }
       setIsPlaying(true);
-      rafRef.current = requestAnimationFrame(step);
       return;
     }
     if (audioContext && audioSource && audioContext.state === "running") {
       audioContext.suspend();
       setIsPlaying(false);
-      rafRef.current && cancelAnimationFrame(rafRef.current);
       return;
     }
   };
@@ -127,21 +123,6 @@ const MusicPlayer = () => {
     };
     loadSong();
   }, [playingIndex]);
-
-  const step = (timestamp: number) => {
-    if (!lastTime) lastTime = timestamp;
-    const progress = timestamp - lastTime;
-    if (progress === 0 || progress > 0) {
-      try {
-        analyserNode?.getByteFrequencyData(visualArr!);
-        setAudioData(Array.from(visualArr!));
-        lastTime = timestamp;
-      } catch (err) {
-        console.log(err);
-      }
-    }
-    rafRef.current = requestAnimationFrame(step);
-  };
 
   return (
     <section className="fixed bottom-[72px] h-[78px] w-full transition-all duration-200 ease-in-out md:bottom-0 md:h-[120px] md:w-[calc(100vw-64px)] lg:w-[calc(100vw-320px)]">
